Move data directory from public to src
