+++ react-app/src/reducers/reducer-active-country.js
@@ -12,7 +12,7 @@ import {
 } from '../helpers/helper-general'
 const config = require('../config.js')
 
-const mpio = require('../../public/data/mpio');
+const mpio = require('../data/mpio');
 // List of scores to represent mobility toward each admin
 // Order is by geoFeature index
 let scores = []
